// ts/src/models/certificates/certificate-enrollments-info-response.ts
export interface CertificateEnrollmentsInfoResponse {
  readonly givenName?: string;
  readonly surname?: string;
  readonly commonName?: string;
  readonly serialNumber?: string;
  readonly countryName?: string;
  readonly uniqueIdentifier?: string;
  readonly organizationName?: string;
  readonly organizationIdentifier?: string;
}